Rebase and rework component
