+++ src/lib/component/partial/KeyValueChip/KeyValueChip.js
@@ -42,7 +42,40 @@ const styles = theme => ({
     border: `1px solid ${emphasize(theme.palette.primary.main, 0.7)}`,
     maxWidth: "200px",
   },
+  imageContainer: {
+    maxWidth: "100%",
+    width: "fit-content",
+  },
+  imageKey: {
+    color: theme.palette.getContrastText(theme.palette.primary.main),
+    background: theme.palette.primary.main,
+    border: `1px solid ${theme.palette.primary.main}`,
+    textAlign: "center",
+    borderRadius: `${theme.spacing(0.5)}px ${theme.spacing(0.5)}px 0 0`,
+    width: "100%",
+    display: "block",
+  },
+  image: {
+    border: `1px solid ${emphasize(theme.palette.primary.main, 0.7)}`,
+    borderRadius: `0 0 ${theme.spacing(0.5)}px ${theme.spacing(0.5)}px`,
+    maxWidth: "100%",
+  },
 });
+
+const imageExtensions = [
+  ".PNG",
+  ".JPG",
+  ".JPEG",
+  ".GIF",
+  ".WEBP",
+  ".TIFF",
+  ".PSD",
+  ".RAW",
+  ".BMP",
+  ".HEIF",
+  ".INDD",
+  ".SVG",
+];
 
 class KeyValueChip extends React.PureComponent {
   static propTypes = {
@@ -55,7 +88,39 @@ class KeyValueChip extends React.PureComponent {
     value: null,
   };
 
-  render() {
+  imageOrStringContainer = () => {
+    const { value } = this.props;
+    try {
+      new URL(value);
+      // see if it's got an image extension
+      let results = imageExtensions.filter(ext =>
+        value.toUpperCase().includes(ext),
+      );
+      let imageMatch = results.length > 0;
+      // if that fails, we should check the content type in case
+      if (!imageMatch) {
+        fetch(value, {
+          method: "HEAD",
+        })
+          .then(response => response.headers.get("Content-type"))
+          .then(type => {
+            imageMatch = type === "image/jpeg" ? true : false;
+          });
+      }
+      // return an image
+      if (imageMatch) {
+        return this.imageContainer();
+      }
+      // return a link
+      return this.stringContainer();
+    } catch (e) {
+      // catch if it's not a url, it's a regular value
+      // We use the same method either way though
+      return this.stringContainer();
+    }
+  };
+
+  stringContainer = () => {
     const { classes, name, value, ...props } = this.props;
     return (
       <Typography component="span" className={classes.root} variant="body2">
@@ -71,11 +136,35 @@ class KeyValueChip extends React.PureComponent {
         </span>
         {value && (
           <span className={classNames(classes.base, classes.value)}>
-            <AutoLink value={value} {...props} />
+            <AutoLink value={value} {...props} />{" "}
           </span>
         )}
       </Typography>
     );
+  };
+
+  imageContainer = () => {
+    const { classes, name, value, ...props } = this.props;
+    return (
+      <Typography component="div" className={classes.root} variant="body2">
+        <div className={classes.imageContainer}>
+          <div
+            className={classNames(classes.base, classes.imageKey)}
+            {...props}
+          >
+            {name}
+          </div>
+          <div {...props}>
+            <img className={this.props.classes.image} src={value} alt={value} />
+          </div>
+        </div>
+      </Typography>
+    );
+  };
+
+  render() {
+    const { classes, name, value, ...props } = this.props;
+    return this.imageOrStringContainer(name, value, classes, props);
   }
 }
